==> Scripts/STEP1_PREPROCESS_SUMMARISE.Rmd <==
---
title: "STEP1_PREPROCESS_SUMMARISE"
subtitle: "MA of transgenerational effects of obesogenic diets on rodents"
author: "ML"
date: "01/05/2020"
output: html_document
knit: (function(inputFile, encoding) {rmarkdown::render(inputFile, encoding = encoding, output_dir = "Knitted") })
---

### Setup
```{r setup, include=FALSE, message=FALSE, warning=FALSE}
rm(list=ls()) # cleaning up objects from the workspace
knitr::opts_chunk$set(echo = FALSE)
options(scipen=100)

# Load packages
# devtools::install_github("thomasp85/patchwork")
pacman::p_load(tidyverse,# tidy family and related pacakges below
               readxl,  #reding in Excel file
               metafor, # calculating effect sizes
               magrittr, # extending piping
               ggplot2,  # nice plots
               patchwork,  # arranging plots
               hrbrthemes # nice themes for ggplots
)

# Load custom functions
#source("../functions.R") #load custom functions

#Load environment

#load('myEnvironment.RData')
```

### Load data    
Unprocessed data for analyses is stored in ALL_TRAITS.xlsx file.   

```{r load and check data, include = FALSE}
#Load Data
#data_all <- read_excel("../data/ALL_TRAITS_20200514.xlsx", sheet = "main")

data_all <- read_excel("../data/ALL_TRAITS_20200514_REVISED.xlsx", sheet = "main")
#Need to import newly revised data
#data_all <- read_excel("../data/ALL_TRAITS.xlsx")
#str(data_all)
# taking a quick look
glimpse(data_all)

# Change spelling in one column
data_all$Exposure_Type <- gsub("One off", "One-off", data_all$Exposure_Type)
#add first colum with ID for each row (to be used as ID for each clculated effect size)
add_column(data_all, ES_ID = paste("ES_",c(1:dim(data_all)[1]),sep=""), .before = 1) -> data_all
# Change all character columns into factors
data_all %>% type.convert() -> data_all 

str(data_all)

#names(data_all)
#colnames(data_all)
```

### Calculating Effect Sizes: lnRR and lnCVR    

*Note for later:* Positive lnRR effect sizes indicate higher mean trait values in grand-offspring of exposed F0 parents, whereas positive lnCVR effect sizes indicate greater trait variance in grand-offspring of F0-exposed parents, relatively to offspring of non-exposed F0 parents. Conversely, negative effect sizes indicate lower mean trait values (lnRR) and lower variance (lnCVR) in grand-offspring of F0-exposed parents, relatively to offspring of non-exposed F0 parents.    

We used escalc function from R package metafor to calculate effect sizes.   

```{r calculate effect sizes, include = FALSE}
# Calculate lnRR and its variance for each row ("ROM" is the log transformed ratio of means (Hedges et al., 1999; Lajeunesse, 2011))
data_all <- escalc(measure="ROM", m1i=Mean_Treatment, sd1i=SD_Treatment, n1i=Sample_Size_n_Treatment, m2i=Mean_Control, sd2i=SD_Control, n2i=Sample_Size_n_Control, var.names=c("lnRR","VlnRR"), data=data_all)

# Calculate lnCVR and its variance for each row ("ROM" is the llog transformed coefficient of variation ratio (akagawa et al., 2015))
data_all <- escalc(measure="CVR", m1i=Mean_Treatment, sd1i=SD_Treatment, n1i=Sample_Size_n_Treatment, m2i=Mean_Control, sd2i=SD_Control, n2i=Sample_Size_n_Control, var.names=c("lnCVR","VlnCVR"), data=data_all)

#str(data_all)
#hist(data_all$lnRR)
#hist(data_all$VlnRR)
#hist(data_all$lnCVR)
#hist(data_all$VlnCVR)

#Check for missing values in each column of the data frame:
colSums(is.na(data_all)) #gives N of missing (NA)

# Save data with effect sizes
write.csv(data_all, file ="../Data/data_all_ES.csv")
save(data_all, file = "../Rdata/data_all_ES.RData")

# Subset by One off and Multigenerational
data_MG <- filter(data_all, Exposure_Type == "Multigenerational")
data_OF <- filter(data_all, Exposure_Type == "One-off")
dim(data_MG)
dim(data_OF)
data_MG <- droplevels(data_MG) #adjust factor levels
data_OF <- droplevels(data_OF) #adjust factor levels
```

Visual check via preliminary funnel plots of all calculated effect sizes and their precisions:

```{r plot effect sizes, include = TRUE}
par(mfrow=c(3,2))
#par(mar=c(4,10,2,0))

## all data
plot(jitter(data_all$lnRR), jitter(sqrt(1/data_all$VlnRR)), cex=0.75,  xlab="lnRR", ylab="Precision (1/SE)", main="all data", xlim=c(-1.1,4.1), ylim=c(0,110)) 
abline(v=0, lwd=0.5)
plot(jitter(data_all$lnCVR), jitter(sqrt(1/data_all$VlnCVR)), cex=0.75,  xlab="lnCVR", ylab="Precision (1/SE)", main="all data", xlim=c(-3,2), ylim=c(1,7)) 
abline(v=0, lwd=0.5)

## OF data
plot(jitter(data_OF$lnRR), jitter(sqrt(1/data_OF$VlnRR)), cex=0.75,  xlab="lnRR", ylab="Precision (1/SE)", main="OF data", xlim=c(-1.1,4.1), ylim=c(0,110))
abline(v=0, lwd=0.5)
plot(jitter(data_OF$lnCVR), jitter(sqrt(1/data_OF$VlnCVR)), cex=0.75,  xlab="lnCVR", ylab="Precision (1/SE)", main="OF data", xlim=c(-3,2), ylim=c(1,7)) 
abline(v=0, lwd=0.5)

## MG data
plot(jitter(data_MG$lnRR), jitter(sqrt(1/data_MG$VlnRR)), cex=0.75,  xlab="lnRR", ylab="Precision (1/SE)", main="MG data", xlim=c(-1.1,4.1), ylim=c(0,110))
abline(v=0, lwd=0.5)
plot(jitter(data_MG$lnCVR), jitter(sqrt(1/data_MG$VlnCVR)), cex=0.75,  xlab="lnCVR", ylab="Precision (1/SE)", main="MG data", xlim=c(-3,2), ylim=c(1,7)) 
abline(v=0, lwd=0.5)
```

```{r check odd effect sizes, include = FALSE}
#filter out two odd points - CHECK!
data_all %>% filter(sqrt(1/data_all$VlnRR)>100) #ES_5 Adedeji2019 F2 males Body_Weight 84 days Table 1 -> checked, correct
data_all %>% filter(lnRR>2 & sqrt(1/data_all$VlnRR)>10) #ES_59 Ding2014 F2 Female Adiposity Parametrial Fat Weight Supp Table and Fig 1b -> checked, correct, but still odd
```

```{r pairwise correlation plots, include = FALSE}
#these plots are quite big, so will not include them in the knitted document

## OF dataset
library(GGally) # for pairwise plots
data_OF %>% ggcorr(palette = "RdBu", label = TRUE) #correlation matrix of all variables 
data_OF %>% select(Treatment_Diet_Energy_kcal_g, Treatment_Diet_Prot_pE, Treatment_Diet_Carb_pE, Treatment_Diet_Fat_pE, Treatment_Diet_PC_ratio, Treatment_Diet_PNP_ratio, Treatment_Start_F0, Treatment_End_F0, Treatment_Duration_F0) %>% ggpairs() #correlations and distributions of treatment-related continous variables 

## MG dataset
library(GGally) # for pairwise plots
data_MG %>% ggcorr(palette = "RdBu", label = TRUE) #correlation matrix of all variables 
data_MG %>% select(Treatment_Diet_Energy_kcal_g, Treatment_Diet_Prot_pE, Treatment_Diet_Carb_pE, Treatment_Diet_Fat_pE, Treatment_Diet_PC_ratio, Treatment_Diet_PNP_ratio, Treatment_Start_F0, Treatment_End_F0, Treatment_Duration_F0) %>% ggpairs() #correlations and distributions of treatment-related continous variables 
```



```{r summarise data, include = FALSE}
# Create quick simple factor tables with numbers of effect sizes per level

#table(data_all$Paper_ID)
#table(data_all$Study_ID)
#table(data_all$Trait)
#table(data_all$Rodent_Type)
#table(data_all$Rodent_Strain)
#table(data_all$Exposure_Type)
#table(data_all$F0_Parent_Exposed)
#table(data_all$Offspring_Generation)
#table(data_all$Offspring_Sex)
#table(data_all$Diet_at_Measurement)
#table(data_all$Lineage_HFD)
#table(data_all$Treatment_Diet_Code)

#hist(data_all$Year)
#hist(data_all$Age_at_Measurement_Days)
#hist(data_all$Treatment_Start_F0)
#hist(data_all$Treatment_End_F0)
#hist(data_all$Treatment_Duration_F0)
#hist(data_all$Treatment_Diet_PC_ratio)
#hist(data_all$Treatment_Diet_PNP_ratio)
#hist(data_all$Treatment_Diet_Fat_pE)
#hist(data_all$Treatment_Diet_Energy_kcal_g)

# Vectors with number of papers per level - all data
N_papers_Trait <- colSums(table(data_all$Paper_ID, data_all$Trait)!=0)
N_papers_Rodent_Type <- colSums(table(data_all$Paper_ID, data_all$Rodent_Type)!=0) 
N_papers_Rodent_Strain <- colSums(table(data_all$Paper_ID, data_all$Rodent_Strain)!=0)
N_papers_Exposure <- colSums(table(data_all$Paper_ID, data_all$Exposure_Type)!=0) 
N_papers_F0_Parent_Exposed <- colSums(table(data_all$Paper_ID, data_all$F0_Parent_Exposed)!=0)
N_papers_Offspring_Generation <- colSums(table(data_all$Paper_ID, data_all$Offspring_Generation)!=0)
N_papers_Offspring_Sex <- colSums(table(data_all$Paper_ID, data_all$Offspring_Sex)!=0)
N_papers_Offspring_Diet <- colSums(table(data_all$Paper_ID, data_all$Diet_at_Measurement)!=0)
N_papers_Lineage_HFD <- colSums(table(data_all$Paper_ID, data_all$Lineage_HFD)!=0)
range(data_all$Treatment_Diet_PC_ratio)
range(data_all$Treatment_Diet_PNP_ratio)
range(data_all$Treatment_Diet_Fat_pE)
range(data_all$Treatment_Diet_Energy_kcal_g)
#colSums(table(data_all$Paper_ID, data_all$Year)!=0) #how many papers by Year

# Vectors with number of papers per level - One-off data
N_papers_Trait_OF <- colSums(table(data_OF$Paper_ID, data_OF$Trait)!=0)
N_papers_Rodent_Type_OF <- colSums(table(data_OF$Paper_ID, data_OF$Rodent_Type)!=0)
N_papers_Rodent_Strain_OF <- colSums(table(data_OF$Paper_ID, data_OF$Rodent_Strain)!=0)
N_papers_Exposure_OF <- colSums(table(data_OF$Paper_ID, data_OF$Exposure_Type)!=0)
N_papers_F0_Parent_Exposed_OF <- colSums(table(data_OF$Paper_ID, data_OF$F0_Parent_Exposed)!=0)
N_papers_Offspring_Generation_OF <- colSums(table(data_OF$Paper_ID, data_OF$Offspring_Generation)!=0)
N_papers_Offspring_Sex_OF <- colSums(table(data_OF$Paper_ID, data_OF$Offspring_Sex)!=0)
N_papers_Offspring_Diet_OF <- colSums(table(data_OF$Paper_ID, data_OF$Diet_at_Measurement)!=0)
N_papers_Lineage_HFD_OF <- colSums(table(data_OF$Paper_ID, data_OF$Lineage_HFD)!=0)
range(data_OF$Treatment_Diet_PC_ratio)
range(data_OF$Treatment_Diet_PNP_ratio)
range(data_OF$Treatment_Diet_Fat_pE)
range(data_OF$Treatment_Diet_Energy_kcal_g)
#colSums(table(data_OF$Paper_ID, data_OF$Year)!=0) #how many papers by Year

# Vectors with number of papers per level - Multigenerational data
N_papers_Trait_MG <- colSums(table(data_MG$Paper_ID, data_MG$Trait)!=0)
N_papers_Rodent_Type_MG <- colSums(table(data_MG$Paper_ID, data_MG$Rodent_Type)!=0)
N_papers_Rodent_Strain_MG <- colSums(table(data_MG$Paper_ID, data_MG$Rodent_Strain)!=0)
N_papers_Exposure_MG <- colSums(table(data_MG$Paper_ID, data_MG$Exposure_Type)!=0) 
N_papers_F0_Parent_Exposed_MG <- colSums(table(data_MG$Paper_ID, data_MG$F0_Parent_Exposed)!=0)
N_papers_Offspring_Generation_MG <- colSums(table(data_MG$Paper_ID, data_MG$Offspring_Generation)!=0)
N_papers_Offspring_Sex_MG <- colSums(table(data_MG$Paper_ID, data_MG$Offspring_Sex)!=0) 
N_papers_Offspring_Diet_MG <- colSums(table(data_MG$Paper_ID, data_MG$Diet_at_Measurement)!=0) 
N_papers_Lineage_HFD_MG <- colSums(table(data_MG$Paper_ID, data_MG$Lineage_HFD)!=0) 
range(data_MG$Treatment_Diet_PC_ratio)
range(data_MG$Treatment_Diet_PNP_ratio)
range(data_MG$Treatment_Diet_Fat_pE)
range(data_MG$Treatment_Diet_Energy_kcal_g)
#colSums(table(data_MG$Paper_ID, data_MG$Year)!=0) #how many papers by Year

## for Figure1
#N_papers_Exposure #N papers
#colSums(table(data_all$Study_ID, data_all$Exposure_Type)!=0) #N studies
#table(data_all$Exposure_Type) #N effect sizes
```


```{r number of individuals}
#sum(na.omit(data_all$Sample_Size_n_Treatment)) # calculate total number of treatment individuals used NOT taking into account multiple measurements
#sum(na.omit(data_all$Sample_Size_n_Control)) # calculate total number of contol individuals in the data_all set if control sharing NOT taken into account and NOT taking into account multiple measurements

N_unique_animals_Treatment <- sum(na.omit(data_all[!duplicated(data_all[,"Cohort_ID"]), "Sample_Size_n_Treatment"])) # calculate total number of treatment individuals used taking into account multiple measurements #1121

#sum(na.omit(data_all[!duplicated(data_all[,"Cohort_ID_Control"]), "Sample_Size_n_Control"])) # calculate total number of contol individuals in the data_all set if control sharing NOT taken into account but taking into account multiple measurements #1371
N_unique_animals_Control <- sum(na.omit(data_all[!duplicated(data_all[ ,"Cohort_ID_Control"]) & !duplicated(data_all[ ,"Cohort_ID"]), "Sample_Size_n_Control"])) # calculate CORRECT total number of contol individuals in the data_all set controlling for control sharing #1060

N_unique_animals <- c(N_unique_animals_Treatment, N_unique_animals_Control)
```


```{r save data for use in RESULTS, include = FALSE}
save(N_papers_Exposure, N_papers_Rodent_Type, N_papers_F0_Parent_Exposed, N_papers_Offspring_Generation, N_papers_Offspring_Sex, N_papers_Trait, N_unique_animals, file = "../Rdata/data_summaries.Rdata")
```



## Back-to-back (pyramid) plots of numbers of ES per factor levels split by Exposure_Type    

```{r plot categories sample sizes - F0_Parent_Exposed, include = TRUE}
#https://bookdown.org/lyzhang10/lzhang_r_tips_book/how-to-plot-data.html
#colors: OF - "#56B4E9", MG - "#E69F00"
table(data_all$F0_Parent_Exposed, data_all$Exposure_Type)

# prepare data for plotting
plotting_df <-
  data_all %>% 
  group_by(F0_Parent_Exposed, Exposure_Type) %>% 
  summarise(Freq = n()) %>% 
  # a trick!
  mutate(Freq = if_else(Exposure_Type == "One-off", -Freq, Freq))

## find the order of F0_Parent_Exposeds by the N
temp_df <-
  plotting_df %>% 
  filter(Exposure_Type == "One-off") %>% 
  arrange(Freq)
the_order <- rev(temp_df$F0_Parent_Exposed)

p_F0_Parent_Exposed <- 
  plotting_df %>% 
  ggplot(aes(x = F0_Parent_Exposed, y = Freq, group = Exposure_Type, fill = Exposure_Type)) +
  geom_bar(stat = "identity", width = 0.5) +
  coord_flip() +
  scale_x_discrete(limits = the_order) + #used ordered F0_Parent_Exposeds - use for single axis
  scale_y_continuous(breaks = seq(-200, 200, 50), 
                     labels = abs(seq(-200, 200, 50)), limits = c(-210, 210)) +
  labs(x = "F0_Parent_Exposed", y = "N effect sizes", title = "Number of effect sizes by F0_Parent_Exposed") +
  theme(legend.position = "bottom",
        legend.title = element_blank(),
        plot.title = element_text(hjust = 0.5),
        panel.background = element_rect(fill =  "white"), 
        panel.grid.major.x = element_line("grey")) +
  scale_fill_manual(values=c("#56B4E9", "#E69F00"),   # change the default colors of bars
                    name="",
                    breaks=c("One-off", "Multigenerational"),
                    labels=c("One-off", "Multigenerational")) 
print(p_F0_Parent_Exposed)
```


```{r plot categories sample sizes - Offspring_Sex, include = TRUE}
#https://bookdown.org/lyzhang10/lzhang_r_tips_book/how-to-plot-data.html
#colors: OF - "#56B4E9", MG - "#E69F00"
table(data_all$Offspring_Sex, data_all$Exposure_Type)

# prepare data for plotting
plotting_df <-
  data_all %>% 
  group_by(Offspring_Sex, Exposure_Type) %>% 
  summarise(Freq = n()) %>% 
  # a trick!
  mutate(Freq = if_else(Exposure_Type == "One-off", -Freq, Freq))

## find the order of Offspring_Sexs by the N
temp_df <-
  plotting_df %>% 
  filter(Exposure_Type == "One-off") %>% 
  arrange(Freq)
the_order <- rev(temp_df$Offspring_Sex)

p_Offspring_Sex <- 
  plotting_df %>% 
  ggplot(aes(x = Offspring_Sex, y = Freq, group = Exposure_Type, fill = Exposure_Type)) +
  geom_bar(stat = "identity", width = 0.5) +
  coord_flip() +
  scale_x_discrete(limits = the_order) + #used ordered Offspring_Sexs - use for single axis
  scale_y_continuous(breaks = seq(-200, 200, 50), 
                     labels = abs(seq(-200, 200, 50)), limits = c(-210, 210)) +
  labs(x = "Offspring_Sex", y = "N effect sizes", title = "Number of effect sizes by Offspring_Sex") +
  theme(legend.position = "bottom",
        legend.title = element_blank(),
        plot.title = element_text(hjust = 0.5),
        panel.background = element_rect(fill =  "white"), 
        panel.grid.major.x = element_line("grey")) +
  scale_fill_manual(values=c("#56B4E9", "#E69F00"),   # change the default colors of bars
                    name="",
                    breaks=c("One-off", "Multigenerational"),
                    labels=c("One-off", "Multigenerational")) 
print(p_Offspring_Sex)
```

```{r plot categories sample sizes - Trait, include = TRUE}
#https://bookdown.org/lyzhang10/lzhang_r_tips_book/how-to-plot-data.html
table(data_all$Trait, data_all$Exposure_Type)

#colors: OF - "#56B4E9", MG - "#E69F00"
# prepare data for plotting
plotting_df <-
  data_all %>% 
  group_by(Trait, Exposure_Type) %>% 
  summarise(Freq = n()) %>% 
  # a trick!
  mutate(Freq = if_else(Exposure_Type == "One-off", -Freq, Freq))

## find the order of Traits by the N
temp_df <-
  plotting_df %>% 
  filter(Exposure_Type == "One-off") %>% 
  arrange(Freq)
the_order <- rev(temp_df$Trait)

p_Trait <- 
  plotting_df %>% 
  ggplot(aes(x = Trait, y = Freq, group = Exposure_Type, fill = Exposure_Type)) +
  geom_bar(stat = "identity", width = 0.5) +
  coord_flip() +
  scale_x_discrete(limits = the_order) + #used ordered Traits - use for single axis
  scale_y_continuous(breaks = seq(-200, 200, 50), 
                     labels = abs(seq(-200, 200, 50)), limits = c(-210, 210)) +
  labs(x = "Trait", y = "N effect sizes", title = "Number of effect sizes by Trait") +
  theme(legend.position = "bottom",
        legend.title = element_blank(),
        plot.title = element_text(hjust = 0.5),
        panel.background = element_rect(fill =  "white"), 
        panel.grid.major.x = element_line("grey")) +
  scale_fill_manual(values=c("#56B4E9", "#E69F00"),   # change the default colors of bars
                    name="",
                    breaks=c("One-off", "Multigenerational"),
                    labels=c("One-off", "Multigenerational")) 
print(p_Trait)
```



*NEXT: run main MA and MR using STEP2_MAIN_ANALYSES.Rmd script*
